Correct event formatting - and move to assert.ts
