+++ test/helpers.ts
@@ -1,12 +1,11 @@
-import { BigNumber, Contract, ContractReceipt, ContractTransaction, utils } from 'ethers';
+import { BigNumber, Contract, utils } from 'ethers';
 import { LogLevel } from '@ethersproject/logger';
 import { PerpMarketConfiguration } from './generated/typechain/MarketConfigurationModule';
 import type { bootstrap } from './bootstrap';
-import { type genTrader, type genOrder, genNumber } from './generators';
+import { type genOrder, type genTrader, genNumber } from './generators';
 import { wei } from '@synthetixio/wei';
 import { fastForwardTo } from '@synthetixio/core-utils/utils/hardhat/rpc';
 import { isNil, uniq } from 'lodash';
-import assert from 'assert';
 
 // --- Constants --- //
 
@@ -162,62 +161,4 @@ export const extendContractAbi = (contract: Contract, abi: string[]) => {
   );
   utils.Logger.setLogLevel(LogLevel.WARNING); // enable default logging again
   return newContract;
-};
-
-const formatDecodedArgs = (value: any): string => {
-  // print nested values as [value1, value2, ...]
-  if (Array.isArray(value)) {
-    return `[${value.map((v) => formatDecodedArgs(v)).join(', ')}]`;
-  }
-
-  // surround string values with quotes
-  if (typeof value === 'string') {
-    return `"${value}"`;
-  }
-
-  return value.toString();
-};
-export const assertEvents = async (
-  tx: ContractTransaction | ContractReceipt,
-  expected: (string | RegExp)[],
-  contract: Contract
-) => {
-  const receipt = 'wait' in tx ? await tx.wait() : tx;
-  const spaces = ' '.repeat(6); // to align with assert output
-
-  const { logs } = receipt;
-  if (logs.length !== expected.length) {
-    throw new Error(`Expected ${expected.length} events, got ${logs.length}`);
-  }
-  let seenEvents: string[] = [];
-  const parsedLogs = logs.map((log, i) => {
-    try {
-      const parsed = contract.interface.parseLog(log);
-      const event = `${parsed.name}(${parsed.args ? formatDecodedArgs(parsed.args) : ''})`;
-      seenEvents.push(event);
-      return event;
-    } catch (error) {
-      throw new Error(
-        `Failed to parse log at index: ${i} \n${spaces}List of parsed events:\n${spaces}${seenEvents.join(
-          `\n${spaces}`
-        )} \n${spaces}Ethers error: ${(error as any).message}`
-      );
-    }
-  });
-  parsedLogs.forEach((event, i) => {
-    const expectedAtIndex = expected[i];
-    if (typeof expectedAtIndex === 'string' ? event === expectedAtIndex : event.match(expectedAtIndex)) {
-      return;
-    } else {
-      const allEvents = parsedLogs.join(`\n${spaces}`);
-
-      typeof expectedAtIndex === 'string'
-        ? assert.strictEqual(
-            event,
-            expectedAtIndex,
-            `Event at index ${i} did not match. \n${spaces}List of parsed events:\n${spaces}${allEvents}`
-          )
-        : assert.match(event, expectedAtIndex);
-    }
-  });
 };
